[UPDATE] Address comments in PR Moving funcs from hooks to regualr funcs
