--- packages/app-shell/src/components/Modal/modals/PlanSelector/hooks/useFeatureFlip.js
@@ -1,7 +1,0 @@
-const useFeatureFlip = (user, featureFlip) => {
-  if (!user || !user.featureFlips) return false;
-
-  return user.featureFlips.includes(featureFlip);
-};
-
-export default useFeatureFlip;
