--- packages/app-shell/src/components/Modal/modals/PlanSelector/hooks/useFilteredListOfPlans.js
@@ -1,9 +1,0 @@
-const useFilteredListOfPlans = (planOptions, planToExclude) => {
-  const planOptionsFiltered = planOptions.filter(
-    (plan) => plan.planId !== planToExclude
-  );
-
-  return planOptionsFiltered;
-};
-
-export default useFilteredListOfPlans;
