+++ packages/app-shell/src/components/Modal/modals/PlanSelector/hooks/useInterval.js
@@ -1,9 +1,9 @@
 import { useState } from 'react';
 
-import useDefaultSelectedPlan from './useDefaultSelectedPlan';
+import { getDefaultSelectedPlan } from '../../../utils';
 
 const useInterval = (planOptions, isUpgradeIntent, user) => {
-  const defaultSelectedPlan = useDefaultSelectedPlan(
+  const defaultSelectedPlan = getDefaultSelectedPlan(
     planOptions,
     user // TODO:REMOVE_WITH_FF:agencyPlan
   );
